refactor: decouple core/services from features/shell new-game dialog (#3546)
Move the new-game leader-selection OpenDialogEvent builder out of
core/services into the shell feature (new_game_leader_dialog_builder.dart)
and inject it at the composition root via
AppEventHandlerScope.extraDialogBuilders, so app_event_handler_scope.dart
no longer imports features/shell. Adds positive (builder constructs the
dialog) and negative (no features/shell import) tests, and documents the
injection seam in SPEC/program/app-ui-wiring.md.

Refs #3546
