+++ app/lib/core/services/app_event_handler_scope.dart
@@ -7,7 +7,6 @@ import 'package:colonizethis_app/package_logger.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:colonizethis_app/app.dart';
-import 'package:colonizethis_app/config/ct_e2e.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy_dialogs.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy_order_helpers.dart';
 import 'package:colonizethis_app/features/game/combat/combat_mode_choice_dialog.dart';
@@ -15,10 +14,7 @@ import 'package:colonizethis_app/features/game/combat/quick_battle_result_dialog
 import 'package:colonizethis_app/features/game/widgets/train_civilians_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/train_military_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/turn_news_dialog.dart';
-import 'package:colonizethis_app/features/shell/new_game_leader_selection_dialog.dart';
-import 'package:colonizethis_app/features/shell/new_game_setup_flow.dart';
 import 'package:colonizethis_app/providers/app_event_bus_provider.dart';
-import 'package:colonizethis_app/providers/blessed_ai_profiles_provider.dart';
 import 'package:colonizethis_app/providers/game_service_provider.dart';
 import 'package:colonizethis_app/providers/observe_session_provider.dart';
 import 'package:colonizethis_app/providers/games_provider.dart';
@@ -56,30 +52,11 @@ const String trainMilitaryDialogId = 'train_military';
 /// [OpenDialogEvent] id for [GrantOrSubsidyDialog]. SPEC/program/app-ui-wiring.md.
 const String grantOrSubsidyDialogId = 'grant_or_subsidy';
 
-/// [OpenDialogEvent] id for [NewGameLeaderSelectionDialog]. SPEC/program/app-ui-wiring.md.
+/// [OpenDialogEvent] id for the new-game leader-selection dialog. The dialog
+/// builder lives in the shell feature (`features/shell/`) and is injected into
+/// this scope at the composition root via [AppEventHandlerScope.extraDialogBuilders]
+/// (Refs #3546). SPEC/program/app-ui-wiring.md.
 const String newGameLeaderSelectionDialogId = 'new_game_leader_selection';
-
-/// Smaller than [GameSetupConfig.defaultConfig]: integration tests compile with
-/// `CT_E2E=true` and must stay inside CI wall clocks (not the locked full-init
-/// 60/30 profile). Production `main` / widget tests use [GameSetupConfig.defaultConfig].
-GameSetupConfig _ctE2eNewGameLeaderTemplateConfig() {
-  final d = GameSetupConfig.defaultConfig;
-  return GameSetupConfig(
-    selectedGreatPowerIds: d.selectedGreatPowerIds,
-    leaderVariantByGpId: d.leaderVariantByGpId,
-    continentCount: 2,
-    minorNationCount: 2,
-    tribeCount: 4,
-    numProvincesOldWorld: 24,
-    numProvincesNewWorld: 12,
-    minProvincesPerMinor: 2,
-    seed: d.seed,
-    infiniteMode: d.infiniteMode,
-    startingResources: d.startingResources,
-    preferredInitialMapZoomMultiplier: d.preferredInitialMapZoomMultiplier,
-    initTownRoadWiringRegionIds: d.initTownRoadWiringRegionIds,
-  );
-}
 
 /// [OpenDialogEvent] id for [CombatModeChoiceDialog]. SPEC/program/app-ui-wiring.md.
 const String combatModeChoiceDialogId = 'combat_mode_choice';
@@ -90,7 +67,6 @@ const String quickBattleResultDialogId = 'quick_battle_result';
 /// [OpenDialogEvent] id for [TurnNewsDialog]. SPEC/program/app-ui-wiring.md.
 const String turnNewsDialogId = 'turn_news';
 
-final _logShell = packageLogger('shell');
 final _logEvent = packageLogger('event');
 
 /// Applies a chosen combat mode to the current game session state.
@@ -188,9 +164,20 @@ Orders _mergeTrainMilitaryOrdersForPlayer({
 /// Binds [AppEventHandler] to [appNavigatorKey] for the app lifetime.
 /// SPEC/program/app-event-bus.md (handler); SPEC/program/app-ui-wiring.md (dialog registration).
 class AppEventHandlerScope extends ConsumerStatefulWidget {
-  const AppEventHandlerScope({super.key, required this.child});
+  const AppEventHandlerScope({
+    super.key,
+    required this.child,
+    this.extraDialogBuilders = const {},
+  });
 
   final Widget child;
+
+  /// Feature-layer dialog builders injected by the composition root, merged
+  /// over the core builders in [_DialogBuilders._dialogBuilders]. This keeps
+  /// `core/services/` free of `features/` dialog imports: a feature owns its
+  /// dialog construction and `main.dart` wires it in by [OpenDialogEvent] id
+  /// (Refs #3546). SPEC/program/app-ui-wiring.md.
+  final Map<String, DialogBuilder> extraDialogBuilders;
 
   @override
   ConsumerState<AppEventHandlerScope> createState() =>
